requirements.txt
click
colorama
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
tensorflow
numpy
soundfile
gunicorn
bs4
python-magic
